requirements.txt
boto3==1.18.45
botocore==1.21.45
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
jmespath==0.10.0
lxml==4.6.3
msgpack==1.0.0
msgpack-numpy==0.4.4.3
numpy==1.21.2
pandas==1.3.3
pikepdf==3.1.0
Pillow==8.3.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.7
indico-client==4.12.6
locust=2.5.1